rename namespace for strava caching protocol
